dashboard/requirements.txt
streamlit>=1.32.0
psycopg2-binary==2.9.9
pandas==2.0.3
plotly==5.18.0
python-dotenv==1.0.1
